+++ src/Screens/Navbar/Navbar.tsx
@@ -3,18 +3,16 @@ import { NavLink } from "react-router-dom";
 import '../index.css';
 import '../css/navbar.css'
 import web from "../images/Capture-removebg-preview.png";
-import "bootstrap/dist/css/bootstrap.min.css";
-const Student_Navbar = () => {
+const Navbar = () => {
     return (
         <>
-       <div className = "container-fluid nav_bg">
+        <div className = "container-fluid nav_bg">
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
-
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
-    <NavLink  className="navbar-brand" to = "/studenthome">
-    <img src={web} className = "image-fluid   logo padl" alt = "home image"/>
+    <NavLink  className="navbar-brand" to = "/">
+    <img src={web} className = "image-fluid  logo" alt = "home image"/>
     </NavLink>
     <div className="collapse navbar-collapse" id="navbarSupportedContent">
       <ul className="navbar-nav me-auto mb-2 mb-lg-0">
@@ -24,7 +22,7 @@ const Student_Navbar = () => {
           exact 
           className="nav-link active" 
           aria-current="page" 
-          to = "/studenthome"
+          to = "/"
           >
             Home
           
@@ -34,26 +32,18 @@ const Student_Navbar = () => {
           <NavLink
           activeClassName = "menu_active" 
           className="nav-link" 
-          to = "/studentcompanies"
+          to = "/about"
           >
-            companies
+            About
+          
           </NavLink>
         </li>
         <li className="nav-item">
           <NavLink
           activeClassName = "menu_active" 
           className="nav-link" 
-          to = "/studentcourses"
-          >
-            Courses
-          </NavLink>
+          to = "/contact">Contact</NavLink>
         </li>
-        <li className="nav-item">
-          <NavLink
-          activeClassName = "menu_active" 
-          className="nav-link" 
-          to = "/logout">Logout</NavLink>
-          </li>
       </ul>
     </div>
   </div>
@@ -65,4 +55,4 @@ const Student_Navbar = () => {
     );
 };
 
-export default Student_Navbar;
+export default Navbar;
