+++ src/Screens/Navbar/Navbar_Show_Applicants.tsx
@@ -4,16 +4,16 @@ import '../index.css';
 import '../css/navbar.css'
 import web from "../images/Capture-removebg-preview.png";
 
-const Navbar_viewcompany = () => {
+const viewcollege_Navbar = () => {
     return (
         <>
        <div className = "container-fluid nav_bg">
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
 
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
-    <NavLink  className="navbar-brand" to = "/adminhome">
+    <NavLink  className="navbar-brand" to = "/">
     <img src={web} className = "image-fluid  logo padl" alt = "home image"/>
     </NavLink>
     <div className="collapse navbar-collapse" id="navbarSupportedContent">
@@ -24,7 +24,7 @@ const Navbar_viewcompany = () => {
           exact 
           className="nav-link active" 
           aria-current="page" 
-          to = "/adminhome"
+          to = "/companyhome"
           >
             Home
           
@@ -34,27 +34,7 @@ const Navbar_viewcompany = () => {
           <NavLink
           activeClassName = "menu_active" 
           className="nav-link" 
-          to = "/adminjobs"
-          >
-            jobs
-          
-          </NavLink>
-        </li>
-        <li className="nav-item">
-          <NavLink
-          activeClassName = "menu_active" 
-          className="nav-link" 
-          to = "/admincompanies"
-          >
-            back
-          
-          </NavLink>
-        </li>
-        <li className="nav-item">
-          <NavLink
-          activeClassName = "menu_active" 
-          className="nav-link" 
-          to = "/logout">Logout</NavLink>
+          to = "/companyjobs">Back</NavLink>
           </li>
       </ul>
     </div>
@@ -67,4 +47,4 @@ const Navbar_viewcompany = () => {
     );
 };
 
-export default Navbar_viewcompany;
+export default viewcollege_Navbar;
