+++ src/Screens/Navbar/Navbar_View_Applicant.tsx
@@ -11,7 +11,7 @@ const viewcollege_Navbar = () => {
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
 
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
     <NavLink  className="navbar-brand" to = "/">
     <img src={web} className = "image-fluid  logo padl" alt = "home image"/>
@@ -22,7 +22,7 @@ const viewcollege_Navbar = () => {
           <NavLink
           activeClassName = "menu_active" 
           className="nav-link" 
-          to = "/companycolleges">Back</NavLink>
+          to = "/companyjobs">Back</NavLink>
           </li>
       </ul>
     </div>
