+++ src/Screens/Navbar/Navbar_View_College.tsx
@@ -10,7 +10,7 @@ const viewcollege_Navbar = () => {
       <div className="container-fluid nav_bg">
         <div className="row navbar">
           <div className="col-10 mx-auto navbar">
-            <nav class="navbar navbar-expand-lg navbar-light">
+            <nav className="navbar navbar-expand-lg navbar-light">
               <div className="container-fluid navbar">
                 <NavLink className="navbar-brand" to="/">
                   <img
