+++ src/Screens/Navbar/Navbar_View_Company.tsx
@@ -11,7 +11,7 @@ const Navbar_viewcompany = () => {
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
 
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
     <NavLink  className="navbar-brand" to = "/adminhome">
     <img src={web} className = "image-fluid  logo padl" alt = "home image"/>
