+++ src/Screens/Navbar/Navbar_View_Std_Company_2.tsx
@@ -3,16 +3,15 @@ import { NavLink } from "react-router-dom";
 import "../index.css";
 import "../css/navbar.css";
 import web from "../images/Capture-removebg-preview.png";
-import "bootstrap/dist/css/bootstrap.min.css";
-const Navbar_Admin = () => {
+const Navbar_viewstdcompany = () => {
   return (
     <>
       <div className="container-fluid nav_bg">
         <div className="row navbar">
           <div className="col-10 mx-auto navbar">
-            <nav class="navbar navbar-expand-lg navbar-light">
+            <nav className="navbar navbar-expand-lg navbar-light">
               <div className="container-fluid navbar">
-                <NavLink className="navbar-brand" to="/">
+                <NavLink className="navbar-brand" to="/studenthome">
                   <img
                     src={web}
                     className="image-fluid  logo padl"
@@ -30,7 +29,7 @@ const Navbar_Admin = () => {
                         exact
                         className="nav-link active"
                         aria-current="page"
-                        to="/adminhome"
+                        to="/studenthome"
                       >
                         Home
                       </NavLink>
@@ -39,7 +38,16 @@ const Navbar_Admin = () => {
                       <NavLink
                         activeClassName="menu_active"
                         className="nav-link"
-                        to="/adminstudent"
+                        to="/studentjobs"
+                      >
+                        jobs
+                      </NavLink>
+                    </li>
+                    <li className="nav-item">
+                      <NavLink
+                        activeClassName="menu_active"
+                        className="nav-link"
+                        to="/viewstudentcompany"
                       >
                         back
                       </NavLink>
@@ -64,4 +72,4 @@ const Navbar_Admin = () => {
   );
 };
 
-export default Navbar_Admin;
+export default Navbar_viewstdcompany;
