+++ src/Screens/View Applicant Student Page/View_Applicant_Student_Page.tsx
