+++ src/Screens/View Company College/View_Company_College.tsx
