+++ src/Screens/View Student Company Page/View_Student_Company_Page.tsx
